requirements.txt
fastapi
uvicorn
python-multipart
numpy
faiss-cpu
requests
pillow
pdf2image
tesserocr
# add other packages you use like pytesseract or sentence-transformers if needed
